% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/br_municipalities.R
\docType{data}
\name{br_municipalities}
\alias{br_municipalities}
\title{A Simple Features Dataframe of Brazilian states and municipalities.}
\format{A data frame with 5564 rows and 4 variables}
\usage{
br_municipalities
}
\description{
\itemize{
  \item name \code{character}. Full name of Brazilian municpality.
  \item state \code{character}. Acronym of Brazilian state.
  \item ibge_code \code{character}. IBGE code of municipality.
  \item geometry \code{character}. Simple features geometry.
}
}
\keyword{datasets}
