% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_municipality}
\alias{select_municipality}
\title{Load Simple Features DataFrame of a Brazilian Municipality}
\usage{
select_municipality(municipality)
}
\arguments{
\item{municipality}{Name of municipality. For a dataframe of these, plus their
state and ID, use \code{data("ibge")}.}
}
\description{
Load Simple Features DataFrame of a Brazilian Municipality
}
