% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_state}
\alias{select_state}
\title{Load Simple Features DataFrame of a Brazilian State}
\usage{
select_state(state)
}
\arguments{
\item{state}{Two-letter abbreviation of Brazilian state name. Use \code{UF()} for
a list.}
}
\description{
Load Simple Features DataFrame of a Brazilian State
}
